+++ src/scripts/manage-dataset/manage-dataset.js
@@ -2263,11 +2263,9 @@ const showCurrentLicense = async () => {
           if (res === "Creative Commons Attribution") {
             $("#button-add-license").hide();
             $("#assign-a-license-header").hide();
-            if ($("#add_license-section").hasClass("is-shown"))
-              document.getElementById("license-assigned").style.display =
-                "block";
 
             licenseContainer.style.display = "block";
+            document.getElementById("license-assigned").style.display = "block";
           } else {
             $("#button-add-license").show();
             $("#assign-a-license-header").show();
@@ -2805,6 +2803,10 @@ $("#button-submit-dataset").click(async () => {
         let didFail = res[2];
         let didUpload = res[3];
         uploadedFolders = res[4];
+
+        // analytics places values with matching action and label pairs into a single 'bucket/aggregate'
+        // so log the increase in size at every step to get the sum total size of the uploaded files
+        incrementInFileSize = uploadedFileSize - previousUploadedFileSize;
 
         // failed to upload a bucket, but did upload some files
         if (didFail && didUpload) {