+++ src/scripts/metadata-files/subjects-samples.js
@@ -967,7 +967,6 @@ async function edit_current_protocol_id(ev) {
   });
 
   if (values) {
-    console.log(values);
     $(currentRow)[0].cells[1].innerHTML =
       "<a href='" + values[0] + "' target='_blank'>" + values[0] + "</a>";
     $(currentRow)[0].cells[2].innerHTML = values[1];