+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Step 4 in Prepare a Dataset. README.txt is now a mandatory metadata file in datasets.
 - Layout of Prepare Metadata tab has been changed. Buttons are spread out more evenly and Manifest file button received an icon change.
+- New landing page and additional tabs in the navigation bar (Documentation and Contact Us)
+- Contact Us tab will provide user information on how to reach out to the team for any issues or suggestions. As well as providing a Gather Logs button for the user to provide in an email for the team to look at any issues in depth.
+- Main login prompt and API login prompt were modified in terms of styling.
+
+## Bug fixes:
+
+- The option to upload duplicate files to an existing Pennsieve dataset through the Organize Datasets feature no longer causes an error while uploading.
+- Hidden files can now be imported except .DS_Store and Thumbs.db files.
 
 ## v.6.0.0 - 2022-05-17